Move l2nloss and nvmask into plugins
